+++ js/jscript.js
@@ -7,12 +7,17 @@ var start = new Date();
 function build_tree() {
 	//задаем параметры дерева
 	var arr = {
-        scrollParent: $("#left_div"),
+        idPrefix:"ftt_",
+        cookieId:"ftt_",
 		extensions: ["persist"],    // расширения куки
+        persist:{
+            expandLazy:true,
+            overrideSource: true
+        },
 		selectMode: 1,
 		generateIds: true,
 		checkbox: false,
-		select: function (e, data) {    //ф-ция выбора нода (селектор, галочка)
+		/*select: function (e, data) {    //ф-ция выбора нода (селектор, галочка)
 			selected_node = "" + $.map(data.tree.getSelectedNodes(), function (node) {  //запоминаем выбранный нод
 				return node.key;
 			});
@@ -21,24 +26,23 @@ function build_tree() {
 				$("#cr_user_cont_sh").attr("value", a);
 				$("#cr_user_cont").attr("value", selected_node);
 			}
-		},
+		},*/
 		toggleEffect: { height: "toggle", duration: 200 },
 		strings: {
 			loading: "Загрузка…",
 			loadError: "Ошибка закгрузки дерева каталогов!"
 		},
 		source : {
-			url: "ajax/ad_ajax_tree.php",
+			url: "index.php",
 			data: {
 				act: "get_tree",
 				type: "folders",
 				pNode: "NULL"
             }
-
 		},
 		lazyLoad: function (e, data) {
 			data.result = $.ajax({
-				url:"ajax/ad_ajax_tree.php",
+				url:"index.php",
 				dataType:"json",
 				data: {
 					act: "get_tree",
@@ -51,16 +55,17 @@ function build_tree() {
 			//$("#tree_objects ul").html('');
 			//$("#tree_objects").fancytree("getRootNode").children =[];
 			var node = data.node;
+            selected_node = node.key;
 			obj_area = $("#tree_objects");
 			obj_area.fancytree(
 				"option",
 				"source",
-				{	url: "ajax/ad_ajax_tree.php",
+				{	url: "index.php",
 					data: {act: 'get_tree', type: 'objects', pNode: node.key},
 					success: function () {}
 				}
 			);
-			//$.getJSON("ajax/ad_ajax_tree.php", {act: 'get_tree', type: 'objects', pNode: node.key}, function (data) {
+			//$.getJSON("index.php", {act: 'get_tree', type: 'objects', pNode: node.key}, function (data) {
 			//		$("#tree_objects").fancytree("getRootNode").addChildren(data);
 			//})
 		}
@@ -69,8 +74,9 @@ function build_tree() {
 }
 
 $(function() {
-	selected_node = "Выберите контейнер в дереве каталогов";
-    $("#dialog_div").dialog({autoOpen:false});
+	selected_node = false;
+    dial_box = $("#dialog_div");
+    dial_box.dialog({autoOpen:false});
     build_tree();
     $("#left_div").resizable({
         handles:"e"
@@ -78,6 +84,8 @@ $(function() {
     //$("#tree_objects").stickyTableHeaders();
 	// строим таблицу объектов контейнера
 	$("#tree_objects").fancytree({
+        idPrefix:"fto_",
+        cookieId:"fto_",
 		//scrollParent: $("#right_div"),
 		extensions: ["persist", "table", "gridnav"],    // расширения куки
 		table: {
@@ -103,8 +111,12 @@ $(function() {
 
 // нажатие кнопки "создать пользователя". запрос формы для ввода.		
 function create_user() {
-    $("#tree").fancytree("option", "checkbox", true);
+    if (!selected_node) {
+        alert("Сначала выберите контейнер!");
+        return false;
+    }
 	//посылаем запрос для получения формы для ввода данных нового пользователя
+    //console.log("selected: " + selected_node);
     $.get(
         "./ajax/ad_create_user.html",
         function (data) {
@@ -118,11 +130,12 @@ function create_user() {
         }
     );
 	// инициируем диалоговое окно с формой
-    $("#dialog_div").dialog({
+    dial_box.dialog({
         title: "Создание нового пользователя",
-        //modal:true,
-	    position: ["center", 100],
+        modal:true,
         width: "600px",
+        position: {my: "center top", at: "center top+10%", of: window},
+        resizable: false,
         buttons: {
             "Создать":function () {
 	            check_cr_user_form();
@@ -138,11 +151,97 @@ function create_user() {
             }
         }
     });
-    $("#dialog_div").dialog("open");
+    dial_box.dialog("open");
 }
-
-
 
 function check_cr_user_form() {
     return false;
+}
+
+function get_locked() {
+    $.get(
+        "./index.php",
+        {act:"get_locked_users"},
+        function (data) {
+            if (data!="false") {
+                var templatediv="<table id='locked_list'>" +
+                    "<thead><tr>" +
+                    "<th></th><th>Пользователь</th><th>Время<br />блокировки</th>" +
+                    "</tr></thead>" +
+                    "<tbody></tbody></div>";
+                $("#dialog_div").html(templatediv);
+                pdata = JSON.parse(data);
+                $("#locked_list").fancytree({
+                    idPrefix:"ftlu_",
+                    cookieId:"ftlu_",
+                    source: pdata,
+                    extensions: ["table", "gridnav"],
+                    table: {
+                        indentation: 20,
+                        nodeColumnIdx: 1,
+                        checkboxColumnIdx: 0
+                    },
+                    gridnav: {
+                        autofocusInput: false,
+                        handleCursorKeys: true
+                    },
+                    imagePath: "./css/img/",
+                    selectMode: 2,
+                    checkbox: true,
+                    strings: {
+                        loading: "Загрузка…",
+                        loadError: "Ошибка закгрузки списка пользователей!"
+                    },
+                    renderColumns: function (e, d) {
+                        var node= d.node,
+                            $tdList = $(node.tr).find(">td");
+                        $tdList.eq(2).text(node.data.tstamp)
+                    }
+                });
+                dial_box.dialog({
+                    title: "Список заблокированных пользователей",
+                    modal:true,
+                    width: "600px",
+                    position: {my: "center top", at: "center top+10%", of: window},
+                    resizable: false,
+                    buttons: {
+                        "Разблокировать":function () {
+                            $("#dialog_div").dialog("close");
+                            s = $.map($("#locked_list").fancytree("getTree").getSelectedNodes(), function(node){
+                                return node.key;
+                            });
+                            unlock_users(s);
+                        },
+                        "Закрыть":function (){
+                            $("#dialog_div").dialog("close");
+                        }
+                    }
+                });
+                dial_box.dialog("open");
+            }
+            else {
+                alert("Нет заблокированых пользователей")
+            }
+        }
+    );
+}
+
+function unlock_users(users) {
+    $.get(
+        "./index.php",
+        {
+            act:"unlock_users",
+            ul:users
+        },
+        function (data) {
+            alert(data)
+        }
+    )
+}
+
+
+function exit() {
+    $("#tree").fancytree("getTree").clearCookies();
+    $("#tree_objects").fancytree("getTree").clearCookies();
+    window.location.replace(window.location + "?exit");
 }